svg/painting/parsing/stroke-width-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="600"><title>SVG Painting: parsing stroke-width with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeWidth" rel="help"/><h:meta content="stroke-width supports only the grammar '&lt;length-percentage&gt;'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("stroke-width", "auto");
test_invalid_value("stroke-width", "10px 20px");
test_invalid_value("stroke-width", "-1px");
test_invalid_value("stroke-width", "-10%");

  ]]></script></svg>
